@@ -19,27 +19,25 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 )
 
-var (
-	// apiRequestDuration tracks the duration separate for each HTTP status
-	// class (1xx, 2xx, ...). This creates a fair amount of time series on
-	// the Prometheus server. Usually, you would track the duration of
-	// serving HTTP request without partitioning by outcome. Do something
-	// like this only if needed. Also note how only status classes are
-	// tracked, not every single status code. The latter would create an
-	// even larger amount of time series. Request counters partitioned by
-	// status code are usually OK as each counter only creates one time
-	// series. Histograms are way more expensive, so partition with care and
-	// only where you really need separate latency tracking. Partitioning by
-	// status class is only an example. In concrete cases, other partitions
-	// might make more sense.
-	apiRequestDuration = prometheus.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Name:    "api_request_duration_seconds",
-			Help:    "Histogram for the request duration of the public API, partitioned by status class.",
-			Buckets: prometheus.ExponentialBuckets(0.1, 1.5, 5),
-		},
-		[]string{"status_class"},
-	)
+// apiRequestDuration tracks the duration separate for each HTTP status
+// class (1xx, 2xx, ...). This creates a fair amount of time series on
+// the Prometheus server. Usually, you would track the duration of
+// serving HTTP request without partitioning by outcome. Do something
+// like this only if needed. Also note how only status classes are
+// tracked, not every single status code. The latter would create an
+// even larger amount of time series. Request counters partitioned by
+// status code are usually OK as each counter only creates one time
+// series. Histograms are way more expensive, so partition with care and
+// only where you really need separate latency tracking. Partitioning by
+// status class is only an example. In concrete cases, other partitions
+// might make more sense.
+var apiRequestDuration = prometheus.NewHistogramVec(
+	prometheus.HistogramOpts{
+		Name:    "api_request_duration_seconds",
+		Help:    "Histogram for the request duration of the public API, partitioned by status class.",
+		Buckets: prometheus.ExponentialBuckets(0.1, 1.5, 5),
+	},
+	[]string{"status_class"},
 )
 
 func handler(w http.ResponseWriter, r *http.Request) {
