@@ -354,7 +354,8 @@ func TestBuckets(t *testing.T) {
 	}
 
 	got = ExponentialBucketsRange(1, 100, 10)
-	want = []float64{1.0, 1.6681005372000588, 2.782559402207125,
+	want = []float64{
+		1.0, 1.6681005372000588, 2.782559402207125,
 		4.641588833612779, 7.742636826811273, 12.915496650148842,
 		21.544346900318846, 35.93813663804629, 59.94842503189414,
 		100.00000000000007,
