@@ -58,7 +58,7 @@ func TestGetOptCodes(t *testing.T) {
 		fmt.Fprintf(w, "%s a[%d:%d], (%s) b[%d:%d] (%s)\n", string(op.Tag),
 			op.I1, op.I2, a[op.I1:op.I2], op.J1, op.J2, b[op.J1:op.J2])
 	}
-	result := string(w.Bytes())
+	result := w.String()
 	expected := `d a[0:1], (q) b[0:0] ()
 e a[1:3], (ab) b[0:2] (ab)
 r a[3:4], (x) b[2:3] (y)
@@ -93,7 +93,7 @@ func TestGroupedOpCodes(t *testing.T) {
 				op.I1, op.I2, op.J1, op.J2)
 		}
 	}
-	result := string(w.Bytes())
+	result := w.String()
 	expected := `group
   e, 5, 8, 5, 8
   i, 8, 8, 8, 9
@@ -134,7 +134,7 @@ four`
 		Context:  3,
 	}
 	result, _ := GetUnifiedDiffString(diff)
-	fmt.Println(strings.Replace(result, "\t", " ", -1))
+	fmt.Println(strings.ReplaceAll(result, "\t", " "))
 	// Output:
 	// --- Original 2005-01-26 23:30:50
 	// +++ Current 2010-04-02 10:20:52
@@ -185,14 +185,14 @@ func TestWithAsciiBJunk(t *testing.T) {
 
 	sm = NewMatcherWithJunk(splitChars(rep("a", 40)+rep("b", 40)),
 		splitChars(rep("a", 44)+rep("b", 40)+rep(" ", 20)), false, isJunk)
-	assertEqual(t, sm.bJunk, map[string]struct{}{" ": struct{}{}})
+	assertEqual(t, sm.bJunk, map[string]struct{}{" ": {}})
 
 	isJunk = func(s string) bool {
 		return s == " " || s == "b"
 	}
 	sm = NewMatcherWithJunk(splitChars(rep("a", 40)+rep("b", 40)),
 		splitChars(rep("a", 44)+rep("b", 40)+rep(" ", 20)), false, isJunk)
-	assertEqual(t, sm.bJunk, map[string]struct{}{" ": struct{}{}, "b": struct{}{}})
+	assertEqual(t, sm.bJunk, map[string]struct{}{" ": {}, "b": {}})
 }
 
 func TestSFBugsRatioForNullSeqn(t *testing.T) {
